Added Exception and FIX ENTITY

+++ demo/demo/src/main/java/com/example/demo/service/AccountService.java
@@ -4,6 +4,7 @@ import com.example.demo.dto.AccountDto;
 import com.example.demo.entity.Account;
 import com.example.demo.entity.User;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface AccountService {
@@ -12,9 +13,9 @@ public interface AccountService {
 
     AccountDto getAccountById(Long id);
 
-    AccountDto deposit(Long id, double amount);
+    AccountDto deposit(Long id, BigDecimal amount);
 
-    AccountDto withdraw(Long id, double amount);
+    AccountDto withdraw(Long id, BigDecimal amount);
 
     List<AccountDto> getAllAccounts();
